Break codegen typing infrastructure out into a common module
